Move local templatetags from resources app to pages
